Move queries/data-fetch to component
